+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Change log
   details: see `Krizalys\Onedrive\Onedrive`
 - `Krizalys\Onedrive\Proxy\DriveItemProxy::getChildren()` now accepts an
   optional `$options` parameter to refine the retrieved result set.
+- `Krizalys\Onedrive\Proxy\DriveItemProxy::upload()`: the `'Content-Type'`
+option is now deprecated & reports a notice; use `'contentType'` instead.
 
 [2.2.0] - 2019-08-07
 --------------------